Clamp timeval math to unsigned int to match progress timestamps. This is fine since only relative differences matter. Closes #3018.

+++ curs_lib.c
@@ -377,7 +377,8 @@ void mutt_progress_init (progress_t* progress, const char *msg,
     dprint (1, (debugfile, "gettimeofday failed: %d\n", errno));
   /* if timestamp is 0 no time-based suppression is done */
   if (TimeInc)
-    progress->timestamp = tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    progress->timestamp = (unsigned int) (tv.tv_sec * 1000)
+        + (unsigned int) (tv.tv_usec / 1000);
   mutt_progress_update (progress, 0, 0);
 }
 
@@ -400,7 +401,8 @@ void mutt_progress_update (progress_t* progress, long pos, int percent)
 
   /* skip refresh if not enough time has passed */
   if (update && progress->timestamp && !gettimeofday (&tv, NULL)) {
-    now = tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    now = (unsigned int) (tv.tv_sec * 1000)
+          + (unsigned int) (tv.tv_usec / 1000);
     if (now && now - progress->timestamp < TimeInc)
       update = 0;
   }
